Move ResultTable to its own file
